Reorganized folder structure and imports. Imports are now relative to files being run in the top directory. testdir is separate from src. Added cmd arg to deactivate a node and all its children.
