+++ src/constants.py
@@ -1,3 +1,3 @@
 LINUX_SLASH = '/'
 WINDOWS_SLASH = '\\'
-FILE_SLASH = LINUX_SLASH
+FILE_SLASH = WINDOWS_SLASH
